transit functions moved into transitfunction file, transit map with transit layer move to its own component
