+++ client/src/components/Home.jsx
@@ -1,40 +1,47 @@
 import { useState, useEffect } from "react";
-import JSZip from 'jszip'
-import Papa from 'papaparse'
-import axios from "axios";
-
 import { GoogleMap, LoadScript } from "@react-google-maps/api";
 import PoiMarkers from "./PoiMarkers";
+import { fetchGTFSData } from "./transitfunction";
+import TransitMap from "./TransitMap";
 
 const Home = () => {
-  const [pointsOfInterest, setPointsOfInterest] = useState([]);
   
-  // This relates to the transit layer state
+  const [pointsOfInterest, setPointsOfInterest] = useState([]);
   const [mapInstance, setMapInstance] = useState(null);
+  const [showTransitMap, setShowTransitMap] = useState(false)
 
   // This is for creating our stops because the transit layer doesn't display them
   const [stops, setStops] = useState([]);
-  
   // This is for creating the shapes, connecting the stops together into routes
   const [shapes, setShapes] = useState([])
   
+
   // This will be triggered on Mapload
   const onMapLoad = (map) => {
     console.log("Map Loaded: ", map)
     setMapInstance(map);
-  
-    // This adds the transit layer when the map is loaded
-    const transitLayer = new google.maps.TransitLayer();
-    transitLayer.setMap(map);
   };
+
+  const mapStyles = [
+    // Turn off points of interest that is default in googlemaps.
+    {
+      featureType: "poi",
+      elementType: "all",
+      stylers: [{ visibility: "off" }],
+    },
+  ];
+
+  //Fetch transit system data
+  useEffect(() => {
+    fetchGTFSData({setStops, setShapes});
+  }, []);
 
   const apiUrl = 
   import.meta.env.MODE === "development"
     ? "http://localhost:8080"
     : import.meta.env.VITE_BACKEND_URL
 
-  console.log("API URL:", apiUrl);
-  
+  //fetching landmarks data from backend
   useEffect(() => {
     fetch(`${apiUrl}/api/landmarks`)
       .then((resp) => {
@@ -48,28 +55,46 @@ const Home = () => {
         setPointsOfInterest(data);
       })
       .catch((err) => console.error("Failed to fetch landmarks", err));
-  }, []);
+  }, [mapInstance]);
+
+  const handleChange = () =>{
+    setMapInstance(null)
+    setShowTransitMap(!showTransitMap)
+  }
 
   return (
-    <LoadScript googleMapsApiKey={import.meta.env.VITE_GOOGLE_MAPS_API_KEY}>
-      <GoogleMap
-        // This is the map component that can be customized
-        mapContainerStyle={{
-          width: "70vh",
-          height: "70vh",
-          marginLeft: "26rem",
-        }}
-        center={{ lat: 38.0406, lng: -84.5037 }}
-        zoom={11.9}
-        mapId="90d6d90b957e9186" // This helps with styling default points of interest
-        gestureHandling={"cooperative"}
-        disableDefaultUI={false}
-        onLoad={onMapLoad}
-      >
-        
-        <PoiMarkers pois={pointsOfInterest}/>
-      </GoogleMap>
-    </LoadScript>
+    <>
+      <button onClick={handleChange}>Transit System</button>
+      <LoadScript googleMapsApiKey={import.meta.env.VITE_GOOGLE_MAPS_API_KEY}>
+        {showTransitMap ? 
+          <TransitMap 
+            mapInstance={mapInstance}
+            setMapInstance={setMapInstance}
+            mapStyles={mapStyles}
+            shapes={shapes}
+            stops={stops}
+          /> 
+          :
+          <GoogleMap
+            // This is the map component that can be customized
+            mapContainerStyle={{
+            width: "70vh",
+            height: "70vh",
+            marginLeft: "26rem",
+            }}
+            center={{ lat: 38.0406, lng: -84.5037 }}
+            zoom={11.9}
+            mapId="90d6d90b957e9186" // This helps with styling default points of interest
+            gestureHandling={"cooperative"}
+            disableDefaultUI={false}
+            onLoad={onMapLoad}
+            options={{ styles: mapStyles }}
+          >
+            <PoiMarkers pois={pointsOfInterest}/>
+          </GoogleMap>
+        }
+      </LoadScript>
+    </>
   );
 };
